Test cleanups (+ optimizations and restructurings)
- Rebase onto warnings PR, streamlined implemented tests
- Replaced suboptimal mersenne twister
- Moved independant types to their own files
- Cleaned out utils.hpp, moved miscellaneous types to `types.hpp`
- No more '_t' names
