+++ tests/speedtest-client.cpp
@@ -147,12 +147,12 @@ int main(int argc, char* argv[])
     std::vector<std::unique_ptr<stream_data>> streams;
     streams.reserve(parallel);
 
-    stream_close_callback_t stream_closed = [&](Stream& s, uint64_t errcode) {
+    stream_close_callback stream_closed = [&](Stream& s, uint64_t errcode) {
         size_t i = s.stream_id >> 2;
         log::critical(test_cat, "Stream {} (rawid={}) closed (error={})", i, s.stream_id, errcode);
     };
 
-    stream_data_callback_t on_stream_data = [&](Stream& s, bstring_view data) {
+    stream_data_callback on_stream_data = [&](Stream& s, bstring_view data) {
         size_t i = s.stream_id >> 2;
         if (i >= parallel)
         {
