+++ tests/speedtest-server.cpp
@@ -76,7 +76,7 @@ int main(int argc, char* argv[])
     auto [listen_addr, listen_port] = parse_addr(server_addr, 5500);
     opt::local_addr server_local{listen_addr, listen_port};
 
-    stream_open_callback_t stream_opened = [&](Stream& s) {
+    stream_open_callback stream_opened = [&](Stream& s) {
         log::warning(test_cat, "Stream {} opened!", s.stream_id);
         return 0;
     };
@@ -95,7 +95,7 @@ int main(int argc, char* argv[])
 
     std::unordered_map<ConnectionID, std::map<int64_t, stream_info>> csd;
 
-    stream_data_callback_t stream_data = [&](Stream& s, bstring_view data) {
+    stream_data_callback stream_data = [&](Stream& s, bstring_view data) {
         auto& sd = csd[s.conn.scid()];
         auto it = sd.find(s.stream_id);
         if (it == sd.end())
